Add threads for dealing with multiple servers
There was a really stupid bug in my previous commit: It cycled through
each connection and waited for some data from the socket. So, if there
was even one boring/empty channel connected, the bot would sit silently
until the heat death of the universe. Now it should do things that are
less stupid.